requirements.txt
numpy
PyOpenGL
pyQt5
heapq
unittest